Replace facility moon lighter to shelter moon lighter, and user security is on

+++ web/WEB-INF/classes/src/org/oscarehr/casemgmt/web/BaseCaseManagementEntryAction.java
@@ -245,7 +245,7 @@ public class BaseCaseManagementEntryAction extends BaseClientAction {
 			else
 				rt = rt + "not major" + "  ";
 			*/
-			if (1==caseStatusId)
+			if (1==caseStatusId.intValue())
 				rt = rt + "resolved";
 			else
 				rt = rt + "unresolved";
